move representation to sdk
